Update UpdateAllProjectiles method
This method now iterates over the projectiles table backwards.
This prevents issues when removing indices.

Every projectile in the table is now also checked to see whether it
is past the screen edges. If it is, that projectile is removed from
the table. This prevents having to iterate through a rapidly growing
table as the game continues.

Updated method comment.